connect.serverClose closes all servers now
The problem was that certain variables were singleton to the module
creating a second server was replacing them and they were not
possible to get a reference to in order to clean it up properly.

This change moves those variables to instance variables on the app
which is stored and then they are all closed when serverClose is
called.